Update d.overall.R
for the "...", any number of individual desirabilities. In the function d.overall() it would be useful to add a check that each individual desirability is a double vector (not matrix). I added a tibble with two columns, which was converted to matrix in line 40. However, since length(list(mymatrix)) == 1, the weights are 1 (rep(1,1)/1). The code runs, but all the resulting desirabilites are NA because of w in line 61.
Good approach for ranking, thank you very much for your work!